report-generator/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.103.2
uvicorn==0.29.0
pika==1.3.2
openai==1.33.0
